[BUGFIX] Make install slightly more robust

+++ install.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
-mv ~/.siege/siege.conf ~/.siege/siege.conf_orig
+mkdir -p ~/.siege
+if [ -f ~/.siege/siege.conf ]
+  mv ~/.siege/siege.conf ~/.siege/siege.conf_orig
+fi
 cp siegerc ~/.siege/siege.conf
 echo "Loadtest siege.conf installed into your homedir. Original stashed away as ~/.siege/siege.conf_orig"
